+++ .github/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Release 0.37.0-dev
 
 ### New features since last release
-* Add `observable` and `measurement` support to `cutensornet` backed `lightning.tensor` C++ layer.
+* Add `observable` and `expval` support to `cutensornet` backed `lightning.tensor` C++ layer.
   [(#728)](https://github.com/PennyLaneAI/pennylane-lightning/pull/728)
 
 * Add gate support to `cutensornet` backed `lightning.tensor` C++ layer.